moved changes to their new locations
